src/test/java/roomescape/repository/ReservationTimeDaoTest.java
package roomescape.repository;

import static org.assertj.core.api.Assertions.assertThat;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.DirtiesContext;
import org.springframework.transaction.annotation.Transactional;
import roomescape.model.ReservationTime;

@Transactional
@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
class ReservationTimeDaoTest {

    @Autowired
    private ReservationTimeDao reservationTimeDao;

    @DisplayName("시간 추가")
    @Test
    void saveReservationTime() {
        ReservationTime time = new ReservationTime(null, "14:00");

        ReservationTime savedTime = reservationTimeDao.save(time);

        assertThat(savedTime.getId()).isEqualTo(1);
    }
}
